+++ packages/fr_acdd/lib/src/enums/fr_acdd_mode.dart
@@ -1,1 +1,1 @@
-enum FrAcddMode { api, bffDto }
+enum FrAcddMode { api, bff }